requirements.txt

reflex==0.7.8a1
sqlalchemy
moment
bcrypt
sqlmodel
twilio
fpdf
psycopg2
python-dotenv
python-multipart
pdf2image